# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HiraNXT Jewellery ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer with purchase history
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetail"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Customer:
      type: object
      required: [id, name, mobile, totalPurchases, balance, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        anniversary:
          type: ["string", "null"]
        totalPurchases:
          type: number
        balance:
          type: number
          description: positive = credit, negative = debit
        loyaltyPoints:
          type: integer
        gstin:
          type: ["string", "null"]
        stateCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, mobile]
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        anniversary:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        stateCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        balance:
          type: number
          description: opening balance — positive = credit (advance), negative = debit (due)
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        birthday:
          type: ["string", "null"]
        anniversary:
          type: ["string", "null"]
        gstin:
          type: ["string", "null"]
        stateCode:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        balance:
          type: number
          description: opening balance, not touched by transactions — positive = credit (advance), negative = debit (due)
    CustomerDetail:
      type: object
      required: [customer, recentSales, pendingRepairs]
      properties:
        customer:
          $ref: "#/components/schemas/Customer"
        recentSales:
          type: array
          items:
            $ref: "#/components/schemas/Sale"
        pendingRepairs:
          type: array
          items:
            $ref: "#/components/schemas/RepairJob"
...
    Sale:
      type: object
      required: [id, customerId, customerName, totalAmount, paidAmount, balanceAmount, gstAmount, discountAmount, exchangeGoldWeight, exchangeGoldValue, paymentMode, paymentStatus, invoiceNumber, saleDate, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceAmount:
          type: number
        gstAmount:
          type: number
        discountAmount:
          type: number
        exchangeGoldWeight:
          type: number
        exchangeGoldValue:
          type: number
        paymentMode:
          type: string
          description: cash, upi, card, bank, credit, partial
        bankAccountId:
          type: ["integer", "null"]
        paymentStatus:
          type: string
          description: paid, partial, pending
        invoiceNumber:
          type: string
        saleDate:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    RepairJob:
      type: object
      required: [id, customerId, customerName, customerMobile, itemDescription, issue, estimatedCost, status, receivedDate, promisedDate, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: string
        customerMobile:
          type: string
        itemDescription:
          type: string
        issue:
          type: string
        estimatedCost:
          type: number
        actualCost:
          type: ["number", "null"]
        status:
          type: string
          description: received, in_progress, ready, delivered
        receivedDate:
          type: string
        promisedDate:
          type: string
        deliveredDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        karigarId:
          type: ["integer", "null"]
        karigarName:
          type: ["string", "null"]
        createdAt:
          type: string
...