# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * HiraNXT Jewellery ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerUpdate {
  name?: string;
  mobile?: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  birthday?: string | null;
  /** @nullable */
  anniversary?: string | null;
  /** @nullable */
  gstin?: string | null;
  /** @nullable */
  stateCode?: string | null;
  /** @nullable */
  notes?: string | null;
  /** opening balance, not touched by transactions — positive = credit (advance), negative = debit (due) */
  balance?: number;
}